refactor(extension-common): Service Worker の配線を移す
サイドパネル/サイドバーの制御、バッジ更新のデバウンス、content script の注入、
広告リンク検証のイベント配線、フレーム CAS の配信、Basic 認証の登録を
setupBackground() として extension-common へ移す。

アプリごとに異なる 3 点は引数で受け取る。

- 警告ページの URL
- バッジに出す件数の数え方 (検証結果の表示形式に依存するため)
- 権限が足りないときの案内ページ

background エントリは設定を渡すだけになった (389 行 → 11 行)。副作用を持つ
cors-basic-auth を含むため ./background のサブパスとして公開し、パッケージの
入口からは読み込まれないようにする。
