+++ apps/inspector/src/components/tabBadge/index.ts
@@ -1,2 +1,1 @@
-export { updateBadge } from "./update-badge";
 export { verifyTabCredentials } from "./verify";
