+++ packages/extension-common/src/tab-badge/update-badge.ts
