+++ packages/extension-common/src/utils/cors-basic-auth.ts
